fastapi_backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
python-dotenv
httpx
PyMuPDF
